pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "praating-ai"
version = "0.3.0"
description = "Push-to-talk voice dictation for Linux using Whisper AI"
readme = "README.md"
authors = [
    {name = "Frederick van Staden"}
]
keywords = ["voice", "dictation", "whisper", "speech-to-text", "transcription"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
requires-python = ">=3.8"
dependencies = [
    "faster-whisper",
    "sounddevice",
    "numpy",
    "pynput",
    "pyyaml",
]

[project.optional-dependencies]
cuda = [
    "nvidia-cudnn-cu12",
    "nvidia-cublas-cu12",
]

[project.urls]
Homepage = "https://github.com/silentFred/praating-ai"
Repository = "https://github.com/silentFred/praating-ai"
Issues = "https://github.com/silentFred/praating-ai/issues"

[project.scripts]
praating = "praating_ai.dictate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
praating_ai = ["sounds/*.wav", "config.yaml"]
